+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 * Fix issue where when reloading a polygon from `localStorage`, the points
   render as white instead of the desired crimson.
 * Allow setting dates with MM/DD/YYYY format or M/D/YYYY
+* After a CMR error, display an error banner, gray out the app; when the
+  user dismisses the banner restore CMR functionality
 
 
 # v0.1.0 (2018-10-16)